Split glci package into multiple files
