+++ internal/glci/glci.go
@@ -6,271 +6,11 @@ import (
 	"fmt"
 
 	"github.com/gardenlinux/glci/internal/cloudprovider"
-	"github.com/gardenlinux/glci/internal/gl"
-	"github.com/gardenlinux/glci/internal/log"
-	"github.com/gardenlinux/glci/internal/ocm"
 )
 
-// Publish publishes a release to all cloud providers specified in the flavors and publishing configurations.
-func Publish(ctx context.Context, flavorsConfig FlavorsConfig, publishingConfig PublishingConfig, aliasesConfig AliasesConfig,
-	creds Credentials, version, commit string,
-) error {
-	ctx = log.WithValues(ctx, "op", "publish", "version", version, "commit", commit)
-
-	log.Debug(ctx, "Loading credentials and configuration")
-	manifestSource, manifestTarget, sources, targets, ocmTarget, err := loadCredentialsAndConfig(ctx, creds, publishingConfig)
-	if err != nil {
-		return fmt.Errorf("invalid credentials or configuration: %w", err)
-	}
-	defer func() {
-		_ = closeSourcesAndTargets(sources, targets, ocmTarget)
-	}()
-
-	publications := make([]cloudprovider.Publication, 0, len(flavorsConfig.Flavors)*2)
-	pubMap := make(map[string][]int, len(flavorsConfig.Flavors))
-	for _, flavor := range flavorsConfig.Flavors {
-		flavorPubs := pubMap[flavor.Cname]
-		found := false
-
-		for _, target := range targets {
-			if target.Type() != flavor.Platform {
-				continue
-			}
-			found = true
-			manifestKey := fmt.Sprintf("meta/singles/%s-%s-%.8s", flavor.Cname, version, commit)
-			lctx := log.WithValues(ctx, "cname", flavor.Cname, "platform", flavor.Platform)
-
-			log.Info(lctx, "Retrieving manifest")
-			var manifest *gl.Manifest
-			manifest, err = cloudprovider.GetManifest(lctx, manifestSource, manifestKey)
-			if err != nil {
-				return fmt.Errorf("cannot get manifest for %s: %w", flavor.Cname, err)
-			}
-			if manifest.Version != version {
-				return fmt.Errorf("manifest for %s has incorrect version %s", flavor.Cname, manifest.Version)
-			}
-			if manifest.BuildCommittish != commit && fmt.Sprintf("%.8s", manifest.BuildCommittish) != commit {
-				return fmt.Errorf("manifest for %s has incorrect commit %s", flavor.Cname, manifest.BuildCommittish)
-			}
-			commit = manifest.BuildCommittish
-
-			log.Debug(lctx, "Retrieving target manifest")
-			var targetManifest *gl.Manifest
-			targetManifest, err = cloudprovider.GetManifest(lctx, manifestTarget, manifestKey)
-			if err != nil && !errors.As(err, &cloudprovider.KeyNotFoundError{}) {
-				return fmt.Errorf("cannot get target manifest for %s: %w", flavor.Cname, err)
-			}
-			if targetManifest != nil {
-				if targetManifest.Version != version {
-					return fmt.Errorf("target manifest for %s has incorrect version %s", flavor.Cname, targetManifest.Version)
-				}
-				if targetManifest.BuildCommittish != commit {
-					return fmt.Errorf("target manifest for %s has incorrect commit %s", flavor.Cname, targetManifest.BuildCommittish)
-				}
-
-				var isPublished bool
-				isPublished, err = target.IsPublished(targetManifest)
-				if err != nil {
-					return fmt.Errorf("cannot determine publishing status for %s: %w", flavor.Cname, err)
-				}
-				if isPublished {
-					log.Info(lctx, "Already published, skipping")
-					continue
-				}
-			}
-
-			publications = append(publications, cloudprovider.Publication{
-				Cname:    flavor.Cname,
-				Manifest: manifest,
-				Target:   target,
-			})
-			pubMap[flavor.Cname] = append(flavorPubs, len(publications)-1)
-		}
-
-		if !found {
-			return fmt.Errorf("no publishing target for %s", flavor.Cname)
-		}
-	}
-
-	var descriptor *ocm.ComponentDescriptor
-	descriptor, err = ocm.BuildComponentDescriptor(ctx, manifestSource, publications, ocmTarget, aliasesConfig, version, commit)
-	if err != nil {
-		return fmt.Errorf("cannot build component descriptor: %w", err)
-	}
-
-	if len(publications) > 0 {
-		log.Info(ctx, "Publishing images", "count", len(publications))
-	} else {
-		log.Info(ctx, "Nothing to publish")
-	}
-
-	for i, publication := range publications {
-		lctx := log.WithValues(ctx, "cname", publication.Cname, "platform", publication.Target.Type())
-
-		log.Info(lctx, "Publishing image")
-		var output cloudprovider.PublishingOutput
-		output, err = publication.Target.Publish(lctx, publication.Cname, publication.Manifest, sources)
-		if err != nil {
-			return fmt.Errorf("cannot publish %s to %s: %w", publication.Cname, publication.Target.Type(), err)
-		}
-
-		manifestOutput := publication.Manifest.PublishedImageMetadata
-		manifestOutput, err = publications[i].Target.AddOwnPublishingOutput(manifestOutput, output)
-		if err != nil {
-			return fmt.Errorf("cannot add publishing output for %s: %w", publication.Cname, err)
-		}
-		publication.Manifest.PublishedImageMetadata = manifestOutput
-		glciVer := glciVersion(ctx)
-		if glciVer != "" {
-			publication.Manifest.GLCIVersion = &glciVer
-		}
-
-		log.Info(lctx, "Updating manifest")
-		manifestKey := fmt.Sprintf("meta/singles/%s-%s-%.8s", publication.Cname, version, commit)
-		err = cloudprovider.PutManifest(lctx, manifestTarget, manifestKey, publication.Manifest)
-		if err != nil {
-			return fmt.Errorf("cannot put manifest for %s: %w", publication.Cname, err)
-		}
-	}
-
-	log.Debug(ctx, "Finalizing component descriptor")
-	err = ocm.AddPublicationOutput(descriptor, publications)
-	if err != nil {
-		return fmt.Errorf("cannot add publication output to component descriptor: %w", err)
-	}
-
-	var descriptorYAML []byte
-	descriptorYAML, err = descriptor.ToYAML()
-	if err != nil {
-		return fmt.Errorf("invalid component descriptor: %w", err)
-	}
-
-	log.Info(ctx, "Publishing component descriptor")
-	err = ocmTarget.PublishComponentDescriptor(ctx, version, descriptorYAML)
-	if err != nil {
-		return fmt.Errorf("cannot publish component descriptor: %w", err)
-	}
-
-	log.Debug(ctx, "Closing sources and targets")
-	err = closeSourcesAndTargets(sources, targets, ocmTarget)
-	if err != nil {
-		return fmt.Errorf("cannot close sources and targets: %w", err)
-	}
-
-	log.Info(ctx, "Publishing completed successfully")
-	return nil
-}
-
-// Remove removes a release from all cloud providers specified in the flavors and publishing configurations.
-func Remove(ctx context.Context, flavorsConfig FlavorsConfig, publishingConfig PublishingConfig, creds Credentials, version, commit string,
-) error {
-	ctx = log.WithValues(ctx, "op", "remove", "version", version, "commit", commit)
-
-	log.Debug(ctx, "Loading credentials and configuration")
-	manifestSource, manifestTarget, sources, targets, ocmTarget, err := loadCredentialsAndConfig(ctx, creds, publishingConfig)
-	if err != nil {
-		return fmt.Errorf("invalid credentials or configuration: %w", err)
-	}
-	defer func() {
-		_ = closeSourcesAndTargets(sources, targets, ocmTarget)
-	}()
-
-	publications := make([]cloudprovider.Publication, 0, len(flavorsConfig.Flavors)*2)
-	pubMap := make(map[string][]int, len(flavorsConfig.Flavors))
-	for _, flavor := range flavorsConfig.Flavors {
-		flavorPubs := pubMap[flavor.Cname]
-		found := false
-
-		for _, target := range targets {
-			if target.Type() != flavor.Platform {
-				continue
-			}
-			found = true
-			manifestKey := fmt.Sprintf("meta/singles/%s-%s-%.8s", flavor.Cname, version, commit)
-			lctx := log.WithValues(ctx, "cname", flavor.Cname, "platform", flavor.Platform)
-
-			log.Info(lctx, "Retrieving manifest")
-			var manifest *gl.Manifest
-			manifest, err = cloudprovider.GetManifest(lctx, manifestTarget, manifestKey)
-			if err != nil {
-				if errors.As(err, &cloudprovider.KeyNotFoundError{}) && manifestTarget != manifestSource {
-					log.Debug(lctx, "Manifest not found, skipping")
-					continue
-				}
-				return fmt.Errorf("cannot get manifest for %s: %w", flavor.Cname, err)
-			}
-			if manifest.Version != version {
-				return fmt.Errorf("manifest for %s has incorrect version %s", flavor.Cname, manifest.Version)
-			}
-			if manifest.BuildCommittish != commit && fmt.Sprintf("%.8s", manifest.BuildCommittish) != commit {
-				return fmt.Errorf("manifest for %s has incorrect commit %s", flavor.Cname, manifest.BuildCommittish)
-			}
-			commit = manifest.BuildCommittish
-
-			var isPublished bool
-			isPublished, err = target.IsPublished(manifest)
-			if err != nil {
-				return fmt.Errorf("cannot determine publishing status for %s: %w", flavor.Cname, err)
-			}
-			if !isPublished {
-				log.Debug(lctx, "Already removed, skipping")
-				continue
-			}
-
-			publications = append(publications, cloudprovider.Publication{
-				Cname:    flavor.Cname,
-				Manifest: manifest,
-				Target:   target,
-			})
-			pubMap[flavor.Cname] = append(flavorPubs, len(publications)-1)
-		}
-		if !found {
-			return fmt.Errorf("no publishing target for %s", flavor.Cname)
-		}
-	}
-
-	if len(publications) > 0 {
-		log.Info(ctx, "Removing images", "count", len(publications))
-	} else {
-		log.Info(ctx, "Nothing to remove")
-	}
-
-	for i, publication := range publications {
-		lctx := log.WithValues(ctx, "cname", publication.Cname, "platform", publication.Target.Type())
-
-		log.Info(lctx, "Removing image")
-		err = publication.Target.Remove(lctx, publication.Manifest, sources)
-		if err != nil {
-			return fmt.Errorf("cannot remove %s from %s: %w", publication.Cname, publication.Target.Type(), err)
-		}
-
-		manifestOutput := publication.Manifest.PublishedImageMetadata
-		manifestOutput, err = publications[i].Target.RemoveOwnPublishingOutput(manifestOutput)
-		if err != nil {
-			return fmt.Errorf("cannot remove publishing output for %s: %w", publication.Cname, err)
-		}
-		publication.Manifest.PublishedImageMetadata = manifestOutput
-		glciVer := glciVersion(ctx)
-		if glciVer != "" {
-			publication.Manifest.GLCIVersion = &glciVer
-		}
-
-		log.Info(lctx, "Updating manifest")
-		manifestKey := fmt.Sprintf("meta/singles/%s-%s-%.8s", publication.Cname, version, commit)
-		err = cloudprovider.PutManifest(lctx, manifestTarget, manifestKey, publication.Manifest)
-		if err != nil {
-			return fmt.Errorf("cannot put manifest for %s: %w", publication.Cname, err)
-		}
-	}
-
-	log.Debug(ctx, "Closing sources and targets")
-	err = closeSourcesAndTargets(sources, targets, ocmTarget)
-	if err != nil {
-		return fmt.Errorf("cannot close sources and targets: %w", err)
-	}
-
-	log.Info(ctx, "Removing completed successfully")
-	return nil
+// WithVersion stores the GLCI version string into the context.
+func WithVersion(ctx context.Context, version string) context.Context {
+	return context.WithValue(ctx, ctxkVer{}, version)
 }
 
 func loadCredentialsAndConfig(ctx context.Context, creds Credentials, publishingConfig PublishingConfig) (cloudprovider.ArtifactSource,
@@ -358,4 +98,11 @@ func closeSourcesAndTargets(sources map[string]cloudprovider.ArtifactSource, tar
 	}
 
 	return errors.Join(errs...)
+}
+
+type ctxkVer struct{}
+
+func glciVersion(ctx context.Context) string {
+	ver, _ := ctx.Value(ctxkVer{}).(string) //nolint:revive // An invalid or missing version results in an empty string.
+	return ver
 }
